requirements.txt
typing==3.6.6
ed25519==1.4
Flask==1.0.2
PyOTA==2.0.7
pycryptodome==3.8.0
